Modified params and models storage and behavior.
Models params are now stored as {key, value} objects. Makes it easier to add values and manipulate the array.

Added a simple templating format for URLs and models parameters. When fetching a model, templates will be replaced with values from the parameters of the report. This means the previous report does not have to know of the following report format at all.